+++ src/marks/component.jsx
@@ -9,6 +9,7 @@ import {
   memoize,
   isApplePaySupported,
   supportsPopups as userAgentSupportsPopups,
+  getUserAgent,
 } from "@krakenjs/belter/src";
 import {
   PLATFORM,
@@ -31,11 +32,16 @@ import type {
 import { BUTTON_LAYOUT, BUTTON_FLOW } from "../constants";
 import { determineEligibleFunding, isFundingEligible } from "../funding";
 import {
+  supportsVenmoPopups,
+  isSupportedNativeVenmoBrowser,
+} from "../funding/util";
+import {
   isSupportedNativeBrowser,
-  getVenmoEligibility,
+  getButtonExperiments,
 } from "../zoid/buttons/util";
 
 import { MarksElement } from "./template";
+import { MarksElementRebrand } from "./templateRebrand";
 
 const DEFAULT_HEIGHT = 20;
 
@@ -50,6 +56,7 @@ type MarksProps = {|
   onShippingAddressChange?: OnShippingAddressChange,
   onShippingOptionsChange?: OnShippingOptionsChange,
   displayOnly?: $ReadOnlyArray<$Values<typeof DISPLAY_ONLY_VALUES>>,
+  userAgent: string,
 |};
 
 export type MarksComponent = (MarksProps) => MarksInstance;
@@ -61,6 +68,7 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
     onShippingAddressChange,
     onShippingOptionsChange,
     displayOnly,
+    userAgent = getUserAgent(),
   }: MarksProps = {}): MarksInstance {
     const height = DEFAULT_HEIGHT;
     const fundingEligibility = getFundingEligibility();
@@ -75,7 +83,17 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
       : false;
     const supportsPopups = userAgentSupportsPopups();
     const supportedNativeBrowser = isSupportedNativeBrowser();
-    const experiment = getVenmoEligibility();
+    const experiment = getButtonExperiments();
+    const supportsVenmoPopup = supportsVenmoPopups(
+      experiment,
+      supportsPopups,
+      userAgent
+    );
+    const supportedNativeVenmoBrowser = isSupportedNativeVenmoBrowser(
+      experiment,
+      userAgent
+    );
+
     const hasShippingCallback = Boolean(
       onShippingChange || onShippingAddressChange || onShippingOptionsChange
     );
@@ -95,8 +113,11 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
       hasShippingCallback,
       supportsPopups,
       supportedNativeBrowser,
+      supportsVenmoPopups: supportsVenmoPopup,
+      supportedNativeVenmoBrowser,
       experiment,
       displayOnly,
+      userAgent,
     });
     const env = getEnv();
 
@@ -118,9 +139,12 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
         flow,
         applePaySupport,
         supportsPopups,
+        supportsVenmoPopups: supportsVenmoPopup,
         supportedNativeBrowser,
+        supportedNativeVenmoBrowser,
         experiment,
         displayOnly,
+        userAgent,
       });
     };
 
@@ -130,16 +154,28 @@ export const getMarksComponent: () => MarksComponent = memoize(() => {
           throw new Error(`${fundingSource || "marks"} not eligible`);
         }
 
+        const isRebrandEnabled = experiment?.isPaypalRebrandEnabled;
+
         getElement(container).appendChild(
           (
             <div>
-              <MarksElement
-                fundingEligibility={fundingEligibility}
-                fundingSources={fundingSources}
-                height={height}
-                experiment={experiment}
-                env={env}
-              />
+              {isRebrandEnabled ? (
+                <MarksElementRebrand
+                  fundingEligibility={fundingEligibility}
+                  fundingSources={fundingSources}
+                  height={height}
+                  experiment={experiment}
+                  env={env}
+                />
+              ) : (
+                <MarksElement
+                  fundingEligibility={fundingEligibility}
+                  fundingSources={fundingSources}
+                  height={height}
+                  experiment={experiment}
+                  env={env}
+                />
+              )}
             </div>
           ).render(dom({ doc: document }))
         );